types: Extend the User interface and add the Course interface Extract the CourseCard component to a separate file
