+++ src/app/dashboard/courses/page.tsx
@@ -3,7 +3,8 @@
 import React, { useEffect, useState } from 'react';
 import { useSession } from 'next-auth/react';
 import Link from 'next/link';
-import { BookOpenIcon, ClockIcon, UsersIcon } from '@/components/ui/Icons';
+import { BookOpenIcon } from '@/components/ui/Icons';
+import CourseCard from '@/components/CourseCard';
 
 interface Course {
   _id: string;
@@ -121,54 +122,5 @@ export default function StudentCoursesPage() {
         </div>
       )}
     </div>
-  );
-}
-
-function CourseCard({ course }: { course: Course }) {
-  const totalDuration = course.modules.reduce(
-    (sum, module) => sum + module.duration,
-    0
-  );
-
-  return (
-    <Link href={`/dashboard/courses/${course._id}`}>
-      <div className="bg-white rounded-lg shadow-md hover:shadow-lg transition-shadow duration-200 overflow-hidden cursor-pointer">
-        {course.thumbnail && (
-          <img
-            src={course.thumbnail}
-            alt={course.title}
-            className="w-full h-48 object-cover"
-          />
-        )}
-        <div className="p-6">
-          <h3 className="text-lg font-semibold text-gray-900 mb-2">
-            {course.title}
-          </h3>
-          <p className="text-gray-600 text-sm mb-4 line-clamp-2">
-            {course.description}
-          </p>
-
-          <div className="flex items-center justify-between text-sm text-gray-500 mb-4">
-            <div className="flex items-center">
-              <UsersIcon className="h-4 w-4 mr-1" />
-              {course.instructor.name}
-            </div>
-            <div className="flex items-center">
-              <ClockIcon className="h-4 w-4 mr-1" />
-              {Math.round(totalDuration)} min
-            </div>
-          </div>
-
-          <div className="flex items-center justify-between">
-            <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-indigo-100 text-indigo-800">
-              {course.level}
-            </span>
-            <span className="text-sm text-gray-500">
-              {course.modules.length} modules
-            </span>
-          </div>
-        </div>
-      </div>
-    </Link>
   );
 }
